requirements.txt
numpy>=1.24
scikit-learn>=1.3
opencv-python>=4.8
matplotlib>=3.7
pillow>=10.0
tqdm>=4.65
jupyter>=1.0
ipykernel>=6.0
